Integrate explicit alias regressions into existing tests
